+++ packages/apollo-angular/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### vNEXT
 
+* Introduces `ExtraSubscriptionOptions`. Allow to run `Apollo.subscribe` outside
+  the Zone ([PR #488](https://github.com/apollographql/apollo-angular/pull/488))
+
 ### v1.0.1
 
 * Brings typed variables to `QueryRef`